requirements.txt
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
python-telegram-bot>=21.0
google-genai>=0.1.0
